rename emu_commands > commands
