+++ commands/debug/__init__.py
@@ -1,4 +1,4 @@
-from emu_commands.base import CommandBase, Command, Flag
+from commands.base import CommandBase, Command, Flag
 from py_utils.emu_utils import run, kill, warning, error
 from py_utils.emu_utils import OPENPILOT_PATH
