+++ commands/panda/__init__.py
@@ -1,5 +1,5 @@
 import importlib
-from emu_commands.base import CommandBase, Command
+from commands.base import CommandBase, Command
 from py_utils.emu_utils import run, error
 from py_utils.emu_utils import OPENPILOT_PATH
